+++ src/app/pages/carnet-detail/carnet-detail.component.ts
@@ -5,7 +5,9 @@ import { Address } from 'src/app/core/models/address';
 import { Contact } from 'src/app/core/models/contact';
 import { AddressService } from 'src/app/core/service/addressService';
 import { ContactService } from 'src/app/core/service/contactService';
+import { AddAdressModalComponent } from 'src/app/modals/add-adress-modal/add-adress-modal.component';
 import { ModifierAdressModalComponent } from 'src/app/modals/modifier-adress-modal/modifier-adress-modal.component';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-carnet-detail',
@@ -22,15 +24,18 @@ export class CarnetDetailComponent implements OnInit {
 
   ngOnInit(): void {
     this.id = this.activatedRoute.snapshot.paramMap.get('id');
+    this.getContact()
+    this.getAdresses();
+  }
 
+  getContact() {
     this.contactService.getContact(this.id).subscribe(data => {
       this.contact = data;
     }, err => {
       console.log(err);
     });
-    this.getAdresses();
-
   }
+
   getAdresses() {
     this.adresseService.getAddresss().subscribe(res => {
       this.adresses = res;
@@ -47,28 +52,44 @@ export class CarnetDetailComponent implements OnInit {
 
   }
   modifierAdresse(contactAdresse: Address) {
-
     const dialogRef = this.dialog.open(ModifierAdressModalComponent, {
       data: { contactAdresse }
     });
     dialogRef.afterClosed().subscribe(result => {
+
       this.getAdresses();
     });
-
-
   }
+
   supprimerAdresse(id: number) {
-// confirm
-    this.adresseService.deleteAddress(id).subscribe(data => {
-      console.log(data);
+    Swal.fire({
+      title: 'Voulez vous supprimer cette adresse?',
+      showDenyButton: true,
+      showCancelButton: false,
+      confirmButtonText: `Supprimer`,
+      denyButtonText: `Annuler`,
+    }).then((result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isConfirmed) {
+        this.adresseService.deleteAddress(id).subscribe(data => {
+          this.getAdresses();
+        }, err => {
+          console.log(err);
+
+        });
+        Swal.fire('Adresse supprimée avec succés', '', 'success')
+      }
+    })
+  }
+
+  openDialog(contactId) {
+    const dialogRef = this.dialog.open(AddAdressModalComponent, {
+      data: { contactId }
+    });
+    dialogRef.afterClosed().subscribe(result => {
       this.getAdresses();
-    }, err => {
-      console.log(err);
 
     });
-
-    //ma mchetech y5arrejli fel data null
-
-
   }
 }
+
